requirements.txt
matplotlib
ipympl
python-frontmatter
qgrid
pbx_gs_python_utils
